% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-functions.r
\name{eff.dim}
\alias{eff.dim}
\title{The effective degrees of freedom of a fitted spline}
\usage{
eff.dim(sd = NULL, B, lam = NULL, d = 2, k = NULL, scale = 1)
}
\description{
The effective number of parameters of a fitted spline depends
on the degree of the spline, the number of knots, and the strength
and degree of the prior used to regularise the spline.
}
