% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux-functions.r
\name{mysetseed}
\alias{mysetseed}
\title{A set-seed clone that initialises at a random value, and returns the value}
\usage{
mysetseed(x = NULL)
}
\arguments{
\item{x}{an optional seed to provide}
}
\value{
the value of the seed that has been set
}
\description{
A set-seed clone that initialises at a random value, and returns the value
}
