% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_model_functions.r
\name{norm_prior}
\alias{norm_prior}
\title{A normal prior with a mix of L1 and L2 penalties}
\usage{
norm_prior(par, sdf = 3, l1penal = 0, idx = 1:length(par))
}
\description{
A normal prior with a mix of L1 and L2 penalties
}
