% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_model_functions.r
\name{postFUN}
\alias{postFUN}
\title{Construct a posterior function for optimisation}
\usage{
postFUN(FUNS, master, args)
}
\arguments{
\item{FUNS}{list of functions (log-likelihoods and log-priors)}

\item{master}{logical matrix giving indices of par vector for each function}

\item{args}{List of arguments to each function}
}
\value{
a function that accepts a single parameter vector and
        evaluates the log prior. The data supplied in 'args' is
        "baked in" to the function environment and does not need
        to be resupplied.
}
\description{
Returns a function of a single parameter vector
}
