% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_model_functions.r
\name{spline2_prior}
\alias{spline2_prior}
\title{Another version of a p-spline prior}
\usage{
spline2_prior(par, sigma = 1, scale = rep(1, length(par)), idx = 1:length(par))
}
\description{
Another version of a p-spline prior
}
