% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_model_functions.r
\name{spline_prior}
\alias{spline_prior}
\title{A prior on d'th differences for p-spline models}
\usage{
spline_prior(par, d = 2, lambda = 1, zero = T, scale = 1, idx = 1:length(par))
}
\description{
A prior on d'th differences for p-spline models
}
